refactoring: moved 'last_commit_to_line' to 'Project' class
